Feat: Tag 데이터 타입 변경
정수형 -> 실수형(Double)

+++ src/main/java/cloneproject/Instagram/domain/feed/dto/PostImageTagRequest.java
@@ -24,13 +24,13 @@ public class PostImageTagRequest {
 	@NotNull(message = "게시물 이미지 태그 x 좌표는 필수입니다.")
 	@Min(value = 0, message = "x 좌표는 0 ~ 100 사이로 입력해주세요.")
 	@Max(value = 100, message = "x 좌표는 0 ~ 100 사이로 입력해주세요.")
-	private Integer tagX;
+	private Double tagX;
 
 	@ApiModelProperty(value = "게시물 이미지 태그 y 좌표", example = "50", required = true)
 	@NotNull(message = "게시물 이미지 태그 y 좌표는 필수입니다.")
 	@Min(value = 0, message = "y 좌표는 0 ~ 100 사이로 입력해주세요.")
 	@Max(value = 100, message = "y 좌표는 0 ~ 100 사이로 입력해주세요.")
-	private Integer tagY;
+	private Double tagY;
 
 	@ApiModelProperty(value = "게시물 이미지 태그 사용자 아이디", example = "dlwlrma", required = true)
 	@NotBlank(message = "게시물 이미지 태그 사용자 아이디는 필수입니다.")
